+++ next/src/routeTree.gen.ts
@@ -17,6 +17,8 @@ import { Route as LayoutQuestionnaireQuestionnaireIdLayoutQCodesListsIndexRouteI
 import { Route as LayoutQuestionnaireQuestionnaireIdLayoutQCodesListsNewRouteImport } from './routes/_layout/questionnaire.$questionnaireId/_layout-q/codes-lists/new'
 import { Route as LayoutQuestionnaireQuestionnaireIdLayoutQCodesListsRouteRouteImport } from './routes/_layout/questionnaire.$questionnaireId/_layout-q/codes-lists/route'
 import { Route as LayoutQuestionnaireQuestionnaireIdLayoutQCompositionRouteImport } from './routes/_layout/questionnaire.$questionnaireId/_layout-q/composition'
+import { Route as LayoutQuestionnaireQuestionnaireIdLayoutQDetailsIndexRouteImport } from './routes/_layout/questionnaire.$questionnaireId/_layout-q/details/index'
+import { Route as LayoutQuestionnaireQuestionnaireIdLayoutQDetailsRouteRouteImport } from './routes/_layout/questionnaire.$questionnaireId/_layout-q/details/route'
 import { Route as LayoutQuestionnaireQuestionnaireIdLayoutQDuplicateVariablesRouteImport } from './routes/_layout/questionnaire.$questionnaireId/_layout-q/duplicate-variables'
 import { Route as LayoutQuestionnaireQuestionnaireIdLayoutQIndexRouteImport } from './routes/_layout/questionnaire.$questionnaireId/_layout-q/index'
 import { Route as LayoutQuestionnaireQuestionnaireIdLayoutQMergeRouteImport } from './routes/_layout/questionnaire.$questionnaireId/_layout-q/merge'
@@ -36,6 +38,7 @@ import { Route as LayoutQuestionnaireQuestionnaireIdLayoutQVariablesRouteRouteIm
 import { Route as LayoutQuestionnaireQuestionnaireIdLayoutQVariablesVariableVariableIdRouteImport } from './routes/_layout/questionnaire.$questionnaireId/_layout-q/variables/variable.$variableId'
 import { Route as LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdArticulationRouteImport } from './routes/_layout/questionnaire.$questionnaireId/_layout-q/version.$versionId/articulation'
 import { Route as LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdCodesListsRouteImport } from './routes/_layout/questionnaire.$questionnaireId/_layout-q/version.$versionId/codes-lists'
+import { Route as LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdDetailsRouteImport } from './routes/_layout/questionnaire.$questionnaireId/_layout-q/version.$versionId/details'
 import { Route as LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdIndexRouteImport } from './routes/_layout/questionnaire.$questionnaireId/_layout-q/version.$versionId/index'
 import { Route as LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdMultimodeRouteImport } from './routes/_layout/questionnaire.$questionnaireId/_layout-q/version.$versionId/multimode'
 import { Route as LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdNomenclaturesRouteImport } from './routes/_layout/questionnaire.$questionnaireId/_layout-q/version.$versionId/nomenclatures'
@@ -156,6 +159,12 @@ const LayoutQuestionnaireQuestionnaireIdLayoutQMultimodeRouteRoute =
     path: '/multimode',
     getParentRoute: () => LayoutQuestionnaireQuestionnaireIdLayoutQRoute,
   } as any)
+const LayoutQuestionnaireQuestionnaireIdLayoutQDetailsRouteRoute =
+  LayoutQuestionnaireQuestionnaireIdLayoutQDetailsRouteRouteImport.update({
+    id: '/details',
+    path: '/details',
+    getParentRoute: () => LayoutQuestionnaireQuestionnaireIdLayoutQRoute,
+  } as any)
 const LayoutQuestionnaireQuestionnaireIdLayoutQCodesListsRouteRoute =
   LayoutQuestionnaireQuestionnaireIdLayoutQCodesListsRouteRouteImport.update({
     id: '/codes-lists',
@@ -199,6 +208,13 @@ const LayoutQuestionnaireQuestionnaireIdLayoutQMultimodeIndexRoute =
     path: '/',
     getParentRoute: () =>
       LayoutQuestionnaireQuestionnaireIdLayoutQMultimodeRouteRoute,
+  } as any)
+const LayoutQuestionnaireQuestionnaireIdLayoutQDetailsIndexRoute =
+  LayoutQuestionnaireQuestionnaireIdLayoutQDetailsIndexRouteImport.update({
+    id: '/',
+    path: '/',
+    getParentRoute: () =>
+      LayoutQuestionnaireQuestionnaireIdLayoutQDetailsRouteRoute,
   } as any)
 const LayoutQuestionnaireQuestionnaireIdLayoutQCodesListsIndexRoute =
   LayoutQuestionnaireQuestionnaireIdLayoutQCodesListsIndexRouteImport.update({
@@ -319,6 +335,15 @@ const LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdMultimodeRoute =
         LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdRouteRoute,
     } as any,
   )
+const LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdDetailsRoute =
+  LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdDetailsRouteImport.update(
+    {
+      id: '/details',
+      path: '/details',
+      getParentRoute: () =>
+        LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdRouteRoute,
+    } as any,
+  )
 const LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdCodesListsRoute =
   LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdCodesListsRouteImport.update(
     {
@@ -356,6 +381,7 @@ export interface FileRoutesByFullPath {
   '/questionnaires/': typeof LayoutQuestionnairesIndexRoute
   '/questionnaire/$questionnaireId/articulation': typeof LayoutQuestionnaireQuestionnaireIdLayoutQArticulationRouteRouteWithChildren
   '/questionnaire/$questionnaireId/codes-lists': typeof LayoutQuestionnaireQuestionnaireIdLayoutQCodesListsRouteRouteWithChildren
+  '/questionnaire/$questionnaireId/details': typeof LayoutQuestionnaireQuestionnaireIdLayoutQDetailsRouteRouteWithChildren
   '/questionnaire/$questionnaireId/multimode': typeof LayoutQuestionnaireQuestionnaireIdLayoutQMultimodeRouteRouteWithChildren
   '/questionnaire/$questionnaireId/nomenclatures': typeof LayoutQuestionnaireQuestionnaireIdLayoutQNomenclaturesRouteRouteWithChildren
   '/questionnaire/$questionnaireId/personalization': typeof LayoutQuestionnaireQuestionnaireIdLayoutQPersonalizationRouteRouteWithChildren
@@ -377,6 +403,7 @@ export interface FileRoutesByFullPath {
   '/questionnaire/$questionnaireId/variables/new': typeof LayoutQuestionnaireQuestionnaireIdLayoutQVariablesNewRoute
   '/questionnaire/$questionnaireId/articulation/': typeof LayoutQuestionnaireQuestionnaireIdLayoutQArticulationIndexRoute
   '/questionnaire/$questionnaireId/codes-lists/': typeof LayoutQuestionnaireQuestionnaireIdLayoutQCodesListsIndexRoute
+  '/questionnaire/$questionnaireId/details/': typeof LayoutQuestionnaireQuestionnaireIdLayoutQDetailsIndexRoute
   '/questionnaire/$questionnaireId/multimode/': typeof LayoutQuestionnaireQuestionnaireIdLayoutQMultimodeIndexRoute
   '/questionnaire/$questionnaireId/nomenclatures/': typeof LayoutQuestionnaireQuestionnaireIdLayoutQNomenclaturesIndexRoute
   '/questionnaire/$questionnaireId/personalization/': typeof LayoutQuestionnaireQuestionnaireIdLayoutQPersonalizationIndexRoute
@@ -384,6 +411,7 @@ export interface FileRoutesByFullPath {
   '/questionnaire/$questionnaireId/variables/variable/$variableId': typeof LayoutQuestionnaireQuestionnaireIdLayoutQVariablesVariableVariableIdRoute
   '/questionnaire/$questionnaireId/version/$versionId/articulation': typeof LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdArticulationRoute
   '/questionnaire/$questionnaireId/version/$versionId/codes-lists': typeof LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdCodesListsRoute
+  '/questionnaire/$questionnaireId/version/$versionId/details': typeof LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdDetailsRoute
   '/questionnaire/$questionnaireId/version/$versionId/multimode': typeof LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdMultimodeRoute
   '/questionnaire/$questionnaireId/version/$versionId/nomenclatures': typeof LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdNomenclaturesRoute
   '/questionnaire/$questionnaireId/version/$versionId/variables': typeof LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdVariablesRoute
@@ -410,6 +438,7 @@ export interface FileRoutesByTo {
   '/questionnaire/$questionnaireId/variables/new': typeof LayoutQuestionnaireQuestionnaireIdLayoutQVariablesNewRoute
   '/questionnaire/$questionnaireId/articulation': typeof LayoutQuestionnaireQuestionnaireIdLayoutQArticulationIndexRoute
   '/questionnaire/$questionnaireId/codes-lists': typeof LayoutQuestionnaireQuestionnaireIdLayoutQCodesListsIndexRoute
+  '/questionnaire/$questionnaireId/details': typeof LayoutQuestionnaireQuestionnaireIdLayoutQDetailsIndexRoute
   '/questionnaire/$questionnaireId/multimode': typeof LayoutQuestionnaireQuestionnaireIdLayoutQMultimodeIndexRoute
   '/questionnaire/$questionnaireId/nomenclatures': typeof LayoutQuestionnaireQuestionnaireIdLayoutQNomenclaturesIndexRoute
   '/questionnaire/$questionnaireId/personalization': typeof LayoutQuestionnaireQuestionnaireIdLayoutQPersonalizationIndexRoute
@@ -417,6 +446,7 @@ export interface FileRoutesByTo {
   '/questionnaire/$questionnaireId/variables/variable/$variableId': typeof LayoutQuestionnaireQuestionnaireIdLayoutQVariablesVariableVariableIdRoute
   '/questionnaire/$questionnaireId/version/$versionId/articulation': typeof LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdArticulationRoute
   '/questionnaire/$questionnaireId/version/$versionId/codes-lists': typeof LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdCodesListsRoute
+  '/questionnaire/$questionnaireId/version/$versionId/details': typeof LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdDetailsRoute
   '/questionnaire/$questionnaireId/version/$versionId/multimode': typeof LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdMultimodeRoute
   '/questionnaire/$questionnaireId/version/$versionId/nomenclatures': typeof LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdNomenclaturesRoute
   '/questionnaire/$questionnaireId/version/$versionId/variables': typeof LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdVariablesRoute
@@ -434,6 +464,7 @@ export interface FileRoutesById {
   '/_layout/questionnaire/$questionnaireId/_layout-q': typeof LayoutQuestionnaireQuestionnaireIdLayoutQRouteWithChildren
   '/_layout/questionnaire/$questionnaireId/_layout-q/articulation': typeof LayoutQuestionnaireQuestionnaireIdLayoutQArticulationRouteRouteWithChildren
   '/_layout/questionnaire/$questionnaireId/_layout-q/codes-lists': typeof LayoutQuestionnaireQuestionnaireIdLayoutQCodesListsRouteRouteWithChildren
+  '/_layout/questionnaire/$questionnaireId/_layout-q/details': typeof LayoutQuestionnaireQuestionnaireIdLayoutQDetailsRouteRouteWithChildren
   '/_layout/questionnaire/$questionnaireId/_layout-q/multimode': typeof LayoutQuestionnaireQuestionnaireIdLayoutQMultimodeRouteRouteWithChildren
   '/_layout/questionnaire/$questionnaireId/_layout-q/nomenclatures': typeof LayoutQuestionnaireQuestionnaireIdLayoutQNomenclaturesRouteRouteWithChildren
   '/_layout/questionnaire/$questionnaireId/_layout-q/personalization': typeof LayoutQuestionnaireQuestionnaireIdLayoutQPersonalizationRouteRouteWithChildren
@@ -455,6 +486,7 @@ export interface FileRoutesById {
   '/_layout/questionnaire/$questionnaireId/_layout-q/variables/new': typeof LayoutQuestionnaireQuestionnaireIdLayoutQVariablesNewRoute
   '/_layout/questionnaire/$questionnaireId/_layout-q/articulation/': typeof LayoutQuestionnaireQuestionnaireIdLayoutQArticulationIndexRoute
   '/_layout/questionnaire/$questionnaireId/_layout-q/codes-lists/': typeof LayoutQuestionnaireQuestionnaireIdLayoutQCodesListsIndexRoute
+  '/_layout/questionnaire/$questionnaireId/_layout-q/details/': typeof LayoutQuestionnaireQuestionnaireIdLayoutQDetailsIndexRoute
   '/_layout/questionnaire/$questionnaireId/_layout-q/multimode/': typeof LayoutQuestionnaireQuestionnaireIdLayoutQMultimodeIndexRoute
   '/_layout/questionnaire/$questionnaireId/_layout-q/nomenclatures/': typeof LayoutQuestionnaireQuestionnaireIdLayoutQNomenclaturesIndexRoute
   '/_layout/questionnaire/$questionnaireId/_layout-q/personalization/': typeof LayoutQuestionnaireQuestionnaireIdLayoutQPersonalizationIndexRoute
@@ -462,6 +494,7 @@ export interface FileRoutesById {
   '/_layout/questionnaire/$questionnaireId/_layout-q/variables/variable/$variableId': typeof LayoutQuestionnaireQuestionnaireIdLayoutQVariablesVariableVariableIdRoute
   '/_layout/questionnaire/$questionnaireId/_layout-q/version/$versionId/articulation': typeof LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdArticulationRoute
   '/_layout/questionnaire/$questionnaireId/_layout-q/version/$versionId/codes-lists': typeof LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdCodesListsRoute
+  '/_layout/questionnaire/$questionnaireId/_layout-q/version/$versionId/details': typeof LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdDetailsRoute
   '/_layout/questionnaire/$questionnaireId/_layout-q/version/$versionId/multimode': typeof LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdMultimodeRoute
   '/_layout/questionnaire/$questionnaireId/_layout-q/version/$versionId/nomenclatures': typeof LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdNomenclaturesRoute
   '/_layout/questionnaire/$questionnaireId/_layout-q/version/$versionId/variables': typeof LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdVariablesRoute
@@ -478,6 +511,7 @@ export interface FileRouteTypes {
     | '/questionnaires/'
     | '/questionnaire/$questionnaireId/articulation'
     | '/questionnaire/$questionnaireId/codes-lists'
+    | '/questionnaire/$questionnaireId/details'
     | '/questionnaire/$questionnaireId/multimode'
     | '/questionnaire/$questionnaireId/nomenclatures'
     | '/questionnaire/$questionnaireId/personalization'
@@ -499,6 +533,7 @@ export interface FileRouteTypes {
     | '/questionnaire/$questionnaireId/variables/new'
     | '/questionnaire/$questionnaireId/articulation/'
     | '/questionnaire/$questionnaireId/codes-lists/'
+    | '/questionnaire/$questionnaireId/details/'
     | '/questionnaire/$questionnaireId/multimode/'
     | '/questionnaire/$questionnaireId/nomenclatures/'
     | '/questionnaire/$questionnaireId/personalization/'
@@ -506,6 +541,7 @@ export interface FileRouteTypes {
     | '/questionnaire/$questionnaireId/variables/variable/$variableId'
     | '/questionnaire/$questionnaireId/version/$versionId/articulation'
     | '/questionnaire/$questionnaireId/version/$versionId/codes-lists'
+    | '/questionnaire/$questionnaireId/version/$versionId/details'
     | '/questionnaire/$questionnaireId/version/$versionId/multimode'
     | '/questionnaire/$questionnaireId/version/$versionId/nomenclatures'
     | '/questionnaire/$questionnaireId/version/$versionId/variables'
@@ -532,6 +568,7 @@ export interface FileRouteTypes {
     | '/questionnaire/$questionnaireId/variables/new'
     | '/questionnaire/$questionnaireId/articulation'
     | '/questionnaire/$questionnaireId/codes-lists'
+    | '/questionnaire/$questionnaireId/details'
     | '/questionnaire/$questionnaireId/multimode'
     | '/questionnaire/$questionnaireId/nomenclatures'
     | '/questionnaire/$questionnaireId/personalization'
@@ -539,6 +576,7 @@ export interface FileRouteTypes {
     | '/questionnaire/$questionnaireId/variables/variable/$variableId'
     | '/questionnaire/$questionnaireId/version/$versionId/articulation'
     | '/questionnaire/$questionnaireId/version/$versionId/codes-lists'
+    | '/questionnaire/$questionnaireId/version/$versionId/details'
     | '/questionnaire/$questionnaireId/version/$versionId/multimode'
     | '/questionnaire/$questionnaireId/version/$versionId/nomenclatures'
     | '/questionnaire/$questionnaireId/version/$versionId/variables'
@@ -555,6 +593,7 @@ export interface FileRouteTypes {
     | '/_layout/questionnaire/$questionnaireId/_layout-q'
     | '/_layout/questionnaire/$questionnaireId/_layout-q/articulation'
     | '/_layout/questionnaire/$questionnaireId/_layout-q/codes-lists'
+    | '/_layout/questionnaire/$questionnaireId/_layout-q/details'
     | '/_layout/questionnaire/$questionnaireId/_layout-q/multimode'
     | '/_layout/questionnaire/$questionnaireId/_layout-q/nomenclatures'
     | '/_layout/questionnaire/$questionnaireId/_layout-q/personalization'
@@ -576,6 +615,7 @@ export interface FileRouteTypes {
     | '/_layout/questionnaire/$questionnaireId/_layout-q/variables/new'
     | '/_layout/questionnaire/$questionnaireId/_layout-q/articulation/'
     | '/_layout/questionnaire/$questionnaireId/_layout-q/codes-lists/'
+    | '/_layout/questionnaire/$questionnaireId/_layout-q/details/'
     | '/_layout/questionnaire/$questionnaireId/_layout-q/multimode/'
     | '/_layout/questionnaire/$questionnaireId/_layout-q/nomenclatures/'
     | '/_layout/questionnaire/$questionnaireId/_layout-q/personalization/'
@@ -583,6 +623,7 @@ export interface FileRouteTypes {
     | '/_layout/questionnaire/$questionnaireId/_layout-q/variables/variable/$variableId'
     | '/_layout/questionnaire/$questionnaireId/_layout-q/version/$versionId/articulation'
     | '/_layout/questionnaire/$questionnaireId/_layout-q/version/$versionId/codes-lists'
+    | '/_layout/questionnaire/$questionnaireId/_layout-q/version/$versionId/details'
     | '/_layout/questionnaire/$questionnaireId/_layout-q/version/$versionId/multimode'
     | '/_layout/questionnaire/$questionnaireId/_layout-q/version/$versionId/nomenclatures'
     | '/_layout/questionnaire/$questionnaireId/_layout-q/version/$versionId/variables'
@@ -722,6 +763,13 @@ declare module '@tanstack/react-router' {
       preLoaderRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQMultimodeRouteRouteImport
       parentRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQRoute
     }
+    '/_layout/questionnaire/$questionnaireId/_layout-q/details': {
+      id: '/_layout/questionnaire/$questionnaireId/_layout-q/details'
+      path: '/details'
+      fullPath: '/questionnaire/$questionnaireId/details'
+      preLoaderRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQDetailsRouteRouteImport
+      parentRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQRoute
+    }
     '/_layout/questionnaire/$questionnaireId/_layout-q/codes-lists': {
       id: '/_layout/questionnaire/$questionnaireId/_layout-q/codes-lists'
       path: '/codes-lists'
@@ -763,6 +811,13 @@ declare module '@tanstack/react-router' {
       fullPath: '/questionnaire/$questionnaireId/multimode/'
       preLoaderRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQMultimodeIndexRouteImport
       parentRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQMultimodeRouteRoute
+    }
+    '/_layout/questionnaire/$questionnaireId/_layout-q/details/': {
+      id: '/_layout/questionnaire/$questionnaireId/_layout-q/details/'
+      path: '/'
+      fullPath: '/questionnaire/$questionnaireId/details/'
+      preLoaderRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQDetailsIndexRouteImport
+      parentRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQDetailsRouteRoute
     }
     '/_layout/questionnaire/$questionnaireId/_layout-q/codes-lists/': {
       id: '/_layout/questionnaire/$questionnaireId/_layout-q/codes-lists/'
@@ -869,6 +924,13 @@ declare module '@tanstack/react-router' {
       preLoaderRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdMultimodeRouteImport
       parentRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdRouteRoute
     }
+    '/_layout/questionnaire/$questionnaireId/_layout-q/version/$versionId/details': {
+      id: '/_layout/questionnaire/$questionnaireId/_layout-q/version/$versionId/details'
+      path: '/details'
+      fullPath: '/questionnaire/$questionnaireId/version/$versionId/details'
+      preLoaderRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdDetailsRouteImport
+      parentRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdRouteRoute
+    }
     '/_layout/questionnaire/$questionnaireId/_layout-q/version/$versionId/codes-lists': {
       id: '/_layout/questionnaire/$questionnaireId/_layout-q/version/$versionId/codes-lists'
       path: '/codes-lists'
@@ -946,6 +1008,21 @@ const LayoutQuestionnaireQuestionnaireIdLayoutQCodesListsRouteRouteChildren: Lay
 const LayoutQuestionnaireQuestionnaireIdLayoutQCodesListsRouteRouteWithChildren =
   LayoutQuestionnaireQuestionnaireIdLayoutQCodesListsRouteRoute._addFileChildren(
     LayoutQuestionnaireQuestionnaireIdLayoutQCodesListsRouteRouteChildren,
+  )
+
+interface LayoutQuestionnaireQuestionnaireIdLayoutQDetailsRouteRouteChildren {
+  LayoutQuestionnaireQuestionnaireIdLayoutQDetailsIndexRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQDetailsIndexRoute
+}
+
+const LayoutQuestionnaireQuestionnaireIdLayoutQDetailsRouteRouteChildren: LayoutQuestionnaireQuestionnaireIdLayoutQDetailsRouteRouteChildren =
+  {
+    LayoutQuestionnaireQuestionnaireIdLayoutQDetailsIndexRoute:
+      LayoutQuestionnaireQuestionnaireIdLayoutQDetailsIndexRoute,
+  }
+
+const LayoutQuestionnaireQuestionnaireIdLayoutQDetailsRouteRouteWithChildren =
+  LayoutQuestionnaireQuestionnaireIdLayoutQDetailsRouteRoute._addFileChildren(
+    LayoutQuestionnaireQuestionnaireIdLayoutQDetailsRouteRouteChildren,
   )
 
 interface LayoutQuestionnaireQuestionnaireIdLayoutQMultimodeRouteRouteChildren {
@@ -1026,6 +1103,7 @@ const LayoutQuestionnaireQuestionnaireIdLayoutQVariablesRouteRouteWithChildren =
 interface LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdRouteRouteChildren {
   LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdArticulationRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdArticulationRoute
   LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdCodesListsRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdCodesListsRoute
+  LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdDetailsRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdDetailsRoute
   LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdMultimodeRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdMultimodeRoute
   LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdNomenclaturesRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdNomenclaturesRoute
   LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdVariablesRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdVariablesRoute
@@ -1038,6 +1116,8 @@ const LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdRouteRouteChildre
       LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdArticulationRoute,
     LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdCodesListsRoute:
       LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdCodesListsRoute,
+    LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdDetailsRoute:
+      LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdDetailsRoute,
     LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdMultimodeRoute:
       LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdMultimodeRoute,
     LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdNomenclaturesRoute:
@@ -1056,6 +1136,7 @@ const LayoutQuestionnaireQuestionnaireIdLayoutQVersionVersionIdRouteRouteWithChi
 interface LayoutQuestionnaireQuestionnaireIdLayoutQRouteChildren {
   LayoutQuestionnaireQuestionnaireIdLayoutQArticulationRouteRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQArticulationRouteRouteWithChildren
   LayoutQuestionnaireQuestionnaireIdLayoutQCodesListsRouteRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQCodesListsRouteRouteWithChildren
+  LayoutQuestionnaireQuestionnaireIdLayoutQDetailsRouteRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQDetailsRouteRouteWithChildren
   LayoutQuestionnaireQuestionnaireIdLayoutQMultimodeRouteRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQMultimodeRouteRouteWithChildren
   LayoutQuestionnaireQuestionnaireIdLayoutQNomenclaturesRouteRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQNomenclaturesRouteRouteWithChildren
   LayoutQuestionnaireQuestionnaireIdLayoutQPersonalizationRouteRoute: typeof LayoutQuestionnaireQuestionnaireIdLayoutQPersonalizationRouteRouteWithChildren
@@ -1076,6 +1157,8 @@ const LayoutQuestionnaireQuestionnaireIdLayoutQRouteChildren: LayoutQuestionnair
       LayoutQuestionnaireQuestionnaireIdLayoutQArticulationRouteRouteWithChildren,
     LayoutQuestionnaireQuestionnaireIdLayoutQCodesListsRouteRoute:
       LayoutQuestionnaireQuestionnaireIdLayoutQCodesListsRouteRouteWithChildren,
+    LayoutQuestionnaireQuestionnaireIdLayoutQDetailsRouteRoute:
+      LayoutQuestionnaireQuestionnaireIdLayoutQDetailsRouteRouteWithChildren,
     LayoutQuestionnaireQuestionnaireIdLayoutQMultimodeRouteRoute:
       LayoutQuestionnaireQuestionnaireIdLayoutQMultimodeRouteRouteWithChildren,
     LayoutQuestionnaireQuestionnaireIdLayoutQNomenclaturesRouteRoute: